+++ jplayer/src/main/java/org/jcodec/player/Stepper.java
@@ -48,7 +48,7 @@ public class Stepper {
 
     private RationalLarge pts;
     // private int frameNo;
-    private int[][] target;
+    private byte[][] target;
     private List<Player.Listener> listeners = new ArrayList<Player.Listener>();
     private Timer timer = new Timer();
 
@@ -67,10 +67,10 @@ public class Stepper {
         ao.open(toJavax(af), af.getFrameSize() * (int) af.getFrameRate());
     }
 
-    private int[][] createTarget() {
+    private byte[][] createTarget() {
         Size dim = mi.getDim();
         int sz = 2 * dim.getWidth() * dim.getHeight();
-        return new int[][] { new int[sz], new int[sz], new int[sz] };
+        return new byte[][] { new byte[sz], new byte[sz], new byte[sz] };
     }
 
     private void nextInt() throws IOException {
